[libs] E: Move Binding for recv()
This commits moves the binding for the recv() system calls to the
syscall crate.
