+++ src/libs/syscall/src/sys/socket/bindings/mod.rs
@@ -12,6 +12,7 @@ pub mod getpeername;
 pub mod getsockname;
 pub mod getsockopt;
 pub mod listen;
+pub mod recv;
 pub mod setsockopt;
 pub mod socket;
 pub mod socketpair;
